Move `ColumnSelector` to its own file
